uic_backend/requirements.txt
Django==1.11.29
django-cors-headers==3.0.2
djangorestframework==3.9.4
gunicorn==19.10.0
pytz==2020.1
